Refactored metrics module. Move metrics constants under MetricEvent file. change metric to metricName. Added Kotlin example in ExampleActivity
