--- libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/internal/navigation/metrics/DirectionsMetrics.kt
@@ -1,6 +1,0 @@
-package com.mapbox.services.android.navigation.v5.internal.navigation.metrics
-
-internal object DirectionsMetrics {
-
-    const val ROUTE_RETRIEVAL = "route_retrieval_event"
-}
