+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/internal/navigation/metrics/NavigationDepartEvent.kt
@@ -1,6 +1,7 @@
 package com.mapbox.services.android.navigation.v5.internal.navigation.metrics
 
 import android.annotation.SuppressLint
+import com.mapbox.services.android.navigation.v5.navigation.metrics.NavigationMetrics
 
 @SuppressLint("ParcelCreator")
 internal class NavigationDepartEvent(
