+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/internal/navigation/metrics/NavigationEvent.kt
@@ -76,6 +76,6 @@ internal abstract class NavigationEvent(
 
     override fun toJson(gson: Gson): String = gson.toJson(this)
 
-    override val metric: String
+    override val metricName: String
         get() = getEventName()
 }
