+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/internal/navigation/metrics/NavigationFeedbackEvent.kt
@@ -3,6 +3,7 @@ package com.mapbox.services.android.navigation.v5.internal.navigation.metrics
 import android.annotation.SuppressLint
 import android.location.Location
 import com.mapbox.services.android.navigation.v5.internal.navigation.routeprogress.MetricsRouteProgress
+import com.mapbox.services.android.navigation.v5.navigation.metrics.NavigationMetrics
 
 @SuppressLint("ParcelCreator")
 internal class NavigationFeedbackEvent(
