+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/internal/utils/extensions/MetricEx.kt
@@ -1,13 +1,13 @@
 package com.mapbox.services.android.navigation.v5.internal.utils.extensions
 
 import com.mapbox.android.telemetry.Event
-import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.DirectionsMetrics
 import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.NavigationAppUserTurnstileEvent
-import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.NavigationMetrics
+import com.mapbox.services.android.navigation.v5.navigation.metrics.DirectionsMetrics
 import com.mapbox.services.android.navigation.v5.navigation.metrics.MetricEvent
+import com.mapbox.services.android.navigation.v5.navigation.metrics.NavigationMetrics
 
 internal fun MetricEvent.toTelemetryEvent(): Event? =
-    when (metric) {
+    when (metricName) {
         DirectionsMetrics.ROUTE_RETRIEVAL,
         NavigationMetrics.ARRIVE,
         NavigationMetrics.CANCEL_SESSION,
