+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/navigation/metrics/MapboxMetricsReporter.kt
@@ -67,7 +67,7 @@ object MapboxMetricsReporter : MetricsReporter {
         }
 
         threadWorker.post {
-            metricsObserver?.onMetricUpdated(metricEvent.metric, metricEvent.toJson(gson))
+            metricsObserver?.onMetricUpdated(metricEvent.metricName, metricEvent.toJson(gson))
         }
     }
