+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/navigation/metrics/MetricEvent.kt
@@ -2,8 +2,6 @@ package com.mapbox.services.android.navigation.v5.navigation.metrics
 
 import androidx.annotation.StringDef
 import com.google.gson.Gson
-import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.DirectionsMetrics
-import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.NavigationMetrics
 
 interface MetricEvent {
 
@@ -20,7 +18,21 @@ interface MetricEvent {
     annotation class Metric
 
     @Metric
-    val metric: String
+    val metricName: String
 
     fun toJson(gson: Gson): String
+}
+
+object NavigationMetrics {
+    const val ARRIVE = "navigation.arrive"
+    const val CANCEL_SESSION = "navigation.cancel"
+    const val DEPART = "navigation.depart"
+    const val REROUTE = "navigation.reroute"
+    const val FEEDBACK = "navigation.feedback"
+    const val INITIAL_GPS = "initial_gps_event"
+    const val APP_USER_TURNSTILE = "appUserTurnstile"
+}
+
+object DirectionsMetrics {
+    const val ROUTE_RETRIEVAL = "route_retrieval_event"
 }
