+++ libandroid-navigation/src/test/java/com/mapbox/services/android/navigation/v5/navigation/metrics/MapboxMetricsReporterTest.kt
@@ -5,7 +5,6 @@ import com.google.gson.Gson
 import com.mapbox.android.telemetry.Event
 import com.mapbox.android.telemetry.MapboxTelemetry
 import com.mapbox.navigation.utils.thread.WorkThreadHandler
-import com.mapbox.services.android.navigation.v5.internal.navigation.metrics.NavigationMetrics
 import com.mapbox.services.android.navigation.v5.internal.utils.extensions.toTelemetryEvent
 import io.mockk.mockk
 import io.mockk.verify
@@ -65,7 +64,7 @@ class MapboxMetricsReporterTest {
     }
 
     private class StubNavigationEvent(
-        override val metric: String
+        override val metricName: String
     ) : Event(), MetricEvent {
 
         override fun writeToParcel(dest: Parcel?, flags: Int) {}
