--- libnavigation-base/src/main/java/com/mapbox/navigation/base/internal/metrics/DirectionsMetrics.kt
@@ -1,6 +1,0 @@
-package com.mapbox.navigation.base.internal.metrics
-
-object DirectionsMetrics {
-
-    const val ROUTE_RETRIEVAL = "route_retrieval_event"
-}
