--- libnavigation-base/src/main/java/com/mapbox/navigation/base/internal/metrics/NavigationMetrics.kt
@@ -1,12 +1,0 @@
-package com.mapbox.navigation.base.internal.metrics
-
-object NavigationMetrics {
-
-    const val ARRIVE = "navigation.arrive"
-    const val CANCEL_SESSION = "navigation.cancel"
-    const val DEPART = "navigation.depart"
-    const val REROUTE = "navigation.reroute"
-    const val FEEDBACK = "navigation.feedback"
-    const val INITIAL_GPS = "initial_gps_event"
-    const val APP_USER_TURNSTILE = "appUserTurnstile"
-}
