+++ libnavigation-base/src/main/java/com/mapbox/navigation/base/metrics/MetricsObserver.kt
@@ -10,9 +10,9 @@ interface MetricsObserver {
     /**
      * Called when metric event handled by [MetricsReporter.addEvent]
      *
-     * @param metric metric event name
+     * @param metricName metric event name
      * @param jsonStringData metric data in JSON string representation
      * @since 1.0.0
      */
-    fun onMetricUpdated(@MetricEvent.Metric metric: String, jsonStringData: String)
+    fun onMetricUpdated(@MetricEvent.Metric metricName: String, jsonStringData: String)
 }
