+++ libnavigation-metrics/build.gradle
@@ -25,8 +25,6 @@ dependencies {
         exclude group: "com.mapbox.mapboxsdk", module: "mapbox-android-core"
     }
 
-    implementation dependenciesList.supportAnnotation
-
     //ktlint
     ktlint dependenciesList.ktlint
