+++ libnavigation-metrics/src/main/java/com/mapbox/navigation/metrics/MapboxMetricsReporter.kt
@@ -69,7 +69,7 @@ object MapboxMetricsReporter : MetricsReporter {
         }
 
         threadWorker.post {
-            metricsObserver?.onMetricUpdated(metricEvent.metric, metricEvent.toJson(gson))
+            metricsObserver?.onMetricUpdated(metricEvent.metricName, metricEvent.toJson(gson))
         }
     }
