+++ libnavigation-metrics/src/main/java/com/mapbox/navigation/metrics/internal/utils/extensions/MetricEx.kt
@@ -1,13 +1,13 @@
 package com.mapbox.navigation.metrics.internal.utils.extensions
 
 import com.mapbox.android.telemetry.Event
-import com.mapbox.navigation.base.internal.metrics.DirectionsMetrics
-import com.mapbox.navigation.base.internal.metrics.NavigationMetrics
+import com.mapbox.navigation.base.metrics.DirectionsMetrics
 import com.mapbox.navigation.base.metrics.MetricEvent
+import com.mapbox.navigation.base.metrics.NavigationMetrics
 import com.mapbox.navigation.metrics.internal.NavigationAppUserTurnstileEvent
 
 internal fun MetricEvent.toTelemetryEvent(): Event? =
-    when (metric) {
+    when (metricName) {
         DirectionsMetrics.ROUTE_RETRIEVAL,
         NavigationMetrics.ARRIVE,
         NavigationMetrics.CANCEL_SESSION,
