+++ libnavigation-metrics/src/test/java/com/mapbox/navigation/metrics/MapboxMetricsReporterTest.kt
@@ -4,8 +4,8 @@ import android.os.Parcel
 import com.google.gson.Gson
 import com.mapbox.android.telemetry.Event
 import com.mapbox.android.telemetry.MapboxTelemetry
-import com.mapbox.navigation.base.internal.metrics.NavigationMetrics
 import com.mapbox.navigation.base.metrics.MetricEvent
+import com.mapbox.navigation.base.metrics.NavigationMetrics
 import com.mapbox.navigation.metrics.internal.utils.extensions.toTelemetryEvent
 import com.mapbox.navigation.utils.thread.WorkThreadHandler
 import io.mockk.mockk
@@ -64,7 +64,7 @@ class MapboxMetricsReporterTest {
     }
 
     private class StubNavigationEvent(
-        override val metric: String
+        override val metricName: String
     ) : Event(), MetricEvent {
 
         override fun writeToParcel(dest: Parcel?, flags: Int) {}
